+++ cardano-api/internal/Cardano/Api/Governance/Actions/ProposalProcedure.hs
@@ -14,7 +14,6 @@ module Cardano.Api.Governance.Actions.ProposalProcedure where
 
 import           Cardano.Api.Address
 import           Cardano.Api.Eon.ShelleyBasedEra
-import           Cardano.Api.Eras.Constraints
 import           Cardano.Api.HasTypeProxy
 import           Cardano.Api.Keys.Shelley
 import           Cardano.Api.ProtocolParameters
