+++ cardano-api/internal/Cardano/Api/LedgerState.hs
@@ -85,7 +85,6 @@ import           Cardano.Api.Block
 import           Cardano.Api.Certificate
 import           Cardano.Api.Eon.ShelleyBasedEra
 import           Cardano.Api.Eras.Case
-import           Cardano.Api.Eras.Constraints
 import           Cardano.Api.Error
 import           Cardano.Api.Genesis
 import           Cardano.Api.IO
